refactor(settings): own update time formatting
